Moved template funcs to pkg/templatefuncs
