+++ pkg/templatefuncs/templatefuncs.go
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU General Public License along
 // with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package config
+package templatefuncs
 
 import (
 	"encoding/json"
